fix delete

+++ ser.py
@@ -43,7 +43,8 @@ def output(data):
                     newp = d[0]
                     ser.write(b'request')
                 else:
-                    requests.delete(url=url_status+code[d[0]])
+                    if code[d[0]] != '':
+                        requests.delete(url=url_status+code[d[0]])
                     code[d[0]] = ''
         port[d[0]]=value
     elif data.find('mat')==0:  # 「mat:<Number>」の形式でマトリックスキーの入力を受ける
